+++ android/src/main/java/me/carda/awesome_notifications/Definitions.java
@@ -22,6 +22,7 @@ public interface Definitions {
     String MEDIA_VALID_ASSET = "^asset?:\\/\\/";
     String MEDIA_VALID_RESOURCE = "^resource?:\\/\\/";
 
+    String BADGE_COUNT = "badgeCount";
     String INITIALIZE_DEBUG_MODE = "debug";
     String INITIALIZE_CHANNELS = "initializeChannels";
     String INITIALIZE_DEFAULT_ICON = "defaultIcon";
@@ -64,10 +65,14 @@ public interface Definitions {
 
     String CHANNEL_METHOD_IS_NOTIFICATION_ALLOWED = "isNotificationAllowed";
     String CHANNEL_METHOD_REQUEST_NOTIFICATIONS = "requestNotifications";
+
     String CHANNEL_METHOD_GET_BADGE_COUNT = "getBadgeCount";
     String CHANNEL_METHOD_SET_BADGE_COUNT = "setBadgeCount";
-    String CHANNEL_METHOD_GET_NEXT_DATE = "getNextDate";
+    String CHANNEL_METHOD_INCREMENT_BADGE_COUNT = "incBadgeCount";
+    String CHANNEL_METHOD_DECREMENT_BADGE_COUNT = "decBadgeCount";
     String CHANNEL_METHOD_RESET_BADGE = "resetBadge";
+
+    String CHANNEL_METHOD_GET_NEXT_DATE = "getNextDate";
     String CHANNEL_METHOD_DISMISS_NOTIFICATION = "dismissNotification";
     String CHANNEL_METHOD_CANCEL_NOTIFICATION = "cancelNotification";
     String CHANNEL_METHOD_CANCEL_SCHEDULE = "cancelSchedule";