+++ ios/Classes/lib/Definitions.swift
@@ -62,12 +62,15 @@ enum Definitions {
     static let  CHANNEL_METHOD_SET_NOTIFICATION_CHANNEL = "setNotificationChannel"
     static let  CHANNEL_METHOD_REMOVE_NOTIFICATION_CHANNEL = "removeNotificationChannel"
 
-    static let  CHANNEL_METHOD_IS_NOTIFICATION_ALLOWED = "isNotificationAllowed"
-    static let  CHANNEL_METHOD_REQUEST_NOTIFICATIONS = "requestNotifications"
     static let  CHANNEL_METHOD_GET_BADGE_COUNT = "getBadgeCount"
     static let  CHANNEL_METHOD_SET_BADGE_COUNT = "setBadgeCount"
+    static let  CHANNEL_METHOD_INCREMENT_BADGE_COUNT = "incBadgeCount"
+    static let  CHANNEL_METHOD_DECREMENT_BADGE_COUNT = "decBadgeCount"
     static let  CHANNEL_METHOD_GET_NEXT_DATE = "getNextDate"
     static let  CHANNEL_METHOD_RESET_BADGE = "resetBadge"
+
+    static let  CHANNEL_METHOD_IS_NOTIFICATION_ALLOWED = "isNotificationAllowed"
+    static let  CHANNEL_METHOD_REQUEST_NOTIFICATIONS = "requestNotifications"
     static let  CHANNEL_METHOD_DISMISS_NOTIFICATION = "dismissNotification"
     static let  CHANNEL_METHOD_CANCEL_NOTIFICATION = "cancelNotification"
     static let  CHANNEL_METHOD_CANCEL_SCHEDULE = "cancelSchedule"
@@ -86,6 +89,7 @@ enum Definitions {
     static let  CHANNEL_METHOD_LIST_ALL_SCHEDULES = "listAllSchedules"
 
     static let  DEFAULT_ICON = "defaultIcon"
+    static let  BADGE_COUNT = "badgeCount"
     static let  SELECT_NOTIFICATION = "SELECT_NOTIFICATION"
     static let  NOTIFICATION_BUTTON_ACTION_PREFIX = "ACTION_NOTIFICATION"