+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.6+12]
+* Added showInCompactView property for MediaPlayer buttons
+* Removed channel key from Android Badge methods, because the segregation in channel keys was never used (now is all global)
+* Added increment and decrement badge methods (more performatic)
 ## [0.0.6+11]
 * Fix Android reschedules on startup process (issue #285)
 * Improved Android channels to manage another package channels and convert then to the new standard, using channelKey as hashKey produced from digest channel content